requirements.txt
pystray==0.19.5
Pillow==10.3.0
